\name{coerce-methods}
\docType{methods}
\alias{coerce,matrix,orientation-method}
\alias{coerce,matrix,eulerzyx-method}
\alias{coerce,matrix,eulerzxz-method}
\alias{coerce,array,orientation-method}
\alias{coerce,orientation,rotvector-method}
\alias{coerce,orientation,eulerzyx-method}
\alias{coerce,orientation,quaternion-method}
\alias{coerce,orientation,skewvector-method}
\alias{coerce,orientation,skewmatrix-method}
\alias{coerce,orientation,eulerzxz-method}
\alias{coerce,rotvector,rotmatrix-method}
\alias{coerce,eulerzyx,rotmatrix-method}
\alias{coerce,quaternion,rotmatrix-method}
\alias{coerce,skewvector,quaternion-method}
\alias{coerce,skewvector,rotmatrix-method}
\alias{coerce,skewmatrix,skewvector-method}
\alias{coerce,skewmatrix,rotmatrix-method}
\alias{coerce,eulerzxz,rotmatrix-method}
\title{ Methods for Function coerce in Package `orientlib' }
\description{ Coercion methods are provided between all types of \code{orientation}
objects, and from matrices to the orientation classes.}
\keyword{methods}

