\name{length-methods}
\docType{methods}
\alias{length,rotmatrix-method}
\alias{length,rotvector-method}
\alias{length,eulerzyx-method}
\alias{length,quaternion-method}
\alias{length,skewvector-method}
\alias{length,skewmatrix-method}
\alias{length,eulerzxz-method}
\title{ Length of orientation object }
\description{ The generic \code{length()} function has methods
for \code{orientations}; it counts the number of orientations in the
object.  }
\keyword{methods}

