\name{matrix-methods}
\docType{methods}
\alias{\%*\%,orientation,orientation-method}
\alias{t,orientation-method}
\alias{^,orientation,numeric-method}
\title{ Methods for matrix operations in `orientlib' }
\description{ Methods are defined for matrix multiplication \code{\%*\%}
transposition \code{t()}, and real powers \code{^}.  These operate on the orientations
term by term.}
\keyword{methods}

