\name{mean-methods}
\docType{methods}
\title{Methods for calculating the mean}
\alias{mean-methods}
\alias{mean,ANY-method}
\alias{mean,orientation-method}
\description{The mean function. }
\section{Methods}{\describe{

\item{x = "ANY"}{ the standard mean function }

\item{x = "orientation"}{ find the nearest SO(3) matrix to the mean \code{\link{rotmatrix-class}}
representation of the orientations }
}}
\keyword{methods}
